Working on three-way merge algorithm
Rather than the existing two-way merge that, during conflicts, always
clobbers and favors the newest write, this commit starts working on a
three-way merge that tries to preserve changes if merging is needed.

Most of the changes was to the interface for the API, which wasn't
thought through to accomodate adding more information to requests and
responses. With this commit the web frontend still works and does
two-way merges, and there's a single failing unit test showing that a
basic three-way merge does not work.