web - less janky experience
For text edits, they're quite peculiar and when you start creating a
text box there is an initial create event that I'm now suppressing. Once
you finish editing a Text box we'll send the SET for the whiteboard.
This prevents the jank before where you'd start editing the text box,
send a SET, and on a subsequence GET you'd lose the selection of the
text box.

I've also extended selection-preservation across concurrent edits. I've
been testing this with multiple VirtualBox virtual machines (2 of them)
connected to the whiteboard. As you create and delete concurrently the
three-way merge usually works (sometimes edits are lost), but unlike
before we also try to preserve the selected item across updates.