mxgraph - vendor and try to export images
There's an old way of exporting images [1] that's able to take a mxGraph
XML document and render an image. It works but the colors and background
are wrong.

The way that GraphEditor exports today [2] actually on the browse first
converts the mxGraphModel to a different XML document format [3], and
only then runs the export in [2]. I was hoping for a pure server-side
export method, maybe this will eventually work I'm not sure.

[1]
https://github.com/jgraph/mxgraph/blob/master/java/test/com/mxgraph/test/OldImageExportTest.java

[2]
https://github.com/jgraph/mxgraph/blob/master/java/test/com/mxgraph/test/mxImageExportTest.java

[3]
https://github.com/jgraph/mxgraph/blob/7af5a44/java/examples/com/mxgraph/examples/web/resources/export.html#L121